+++ packages/icestark/src/AppRoute.tsx
@@ -1,12 +1,16 @@
 import * as React from 'react';
 import renderComponent from './util/renderComponent';
 import { AppHistory } from './appHistory';
-import { unloadMicroApp, BaseConfig, createMicroApp } from './apps';
+import { unloadMicroApp, unmountMicroApp, BaseConfig, createMicroApp } from './apps';
 import { converArray2String } from './util/helpers';
-import { PathData } from './util/checkActive';
-import { callCapturedEventListeners, resetCapturedEventListeners } from './util/capturedListeners';
 import { started } from './start';
+import {
+  callCapturedEventListeners,
+  resetCapturedEventListeners,
+} from './util/capturedListeners';
 import isEqual from 'lodash.isequal';
+
+import type { PathData } from './util/checkActive';
 
 interface AppRouteState {
   showComponent: boolean;
@@ -135,6 +139,7 @@ export default class AppRoute extends React.Component<AppRouteProps, AppRouteSta
 
   mountApp = () => {
     resetCapturedEventListeners();
+
     const { onAppEnter } = this.props;
 
     // Trigger app enter
@@ -150,7 +155,7 @@ export default class AppRoute extends React.Component<AppRouteProps, AppRouteSta
   };
 
   unmountApp = () => {
-    const { name, onAppLeave } = this.props;
+    const { name, onAppLeave, cached } = this.props;
 
     // Trigger app leave
     if (typeof onAppLeave === 'function') {
@@ -158,7 +163,7 @@ export default class AppRoute extends React.Component<AppRouteProps, AppRouteSta
     }
 
     if (!this.validateRender() && started) {
-      unloadMicroApp(name);
+      cached ? unmountMicroApp(name) : unloadMicroApp(name);
     }
   };