+++ packages/icestark/src/util/handleAssets.ts
@@ -57,7 +57,7 @@ export interface ILifecycleProps {
   customProps?: object;
 }
 
-function isAssetExist(element: HTMLScriptElement | HTMLLIElement, type: 'script' | 'link') {
+function isAssetExist(element: HTMLScriptElement | HTMLLinkElement, type: 'script' | 'link') {
   const urlAlias = type === 'script' ? 'src' : 'href';
 
   return Array.from(document.getElementsByTagName(type))
@@ -75,11 +75,16 @@ function isAssetExist(element: HTMLScriptElement | HTMLLIElement, type: 'script'
 /**
  * Create link/style element and append to root
  */
-export function appendCSS(
-  root: HTMLElement | ShadowRoot,
-  asset: Asset | HTMLElement,
-  id: string,
-): Promise<void> {
+export function appendCSS(asset: Asset | HTMLElement,
+  {
+    root,
+    id,
+    cacheId,
+  }: {
+    root?: HTMLElement | ShadowRoot;
+    id?: string;
+    cacheId?: string;
+  }): Promise<void> {
   return new Promise<void>(async (resolve, reject) => {
     if (!root) reject(new Error('no root element for css asset'));
 
@@ -95,6 +100,7 @@ export function appendCSS(
       const styleElement: HTMLStyleElement = document.createElement('style');
       styleElement.id = id;
       styleElement.setAttribute(PREFIX, DYNAMIC);
+      cacheId && styleElement.setAttribute('data-cache', cacheId);
       styleElement.innerHTML = content;
       root.appendChild(styleElement);
       resolve();
@@ -112,6 +118,7 @@ export function appendCSS(
         styleElement.innerHTML = await cachedStyleContent[content];
         styleElement.id = id;
         styleElement.setAttribute(PREFIX, DYNAMIC);
+        cacheId && styleElement.setAttribute('data-cache', cacheId);
         root.appendChild(styleElement);
         useExternalLink = false;
         resolve();
@@ -126,6 +133,7 @@ export function appendCSS(
       element.id = id;
       element.rel = 'stylesheet';
       element.href = content;
+      cacheId && element.setAttribute('data-cache', cacheId);
 
       element.addEventListener(
         'error',
@@ -156,18 +164,20 @@ function setAttributeForScriptNode(element: HTMLScriptElement, {
   id,
   src,
   scriptAttributes,
+  cacheId,
 }: {
   module: boolean;
   id: string;
   src: string;
   scriptAttributes: ScriptAttributes;
+  cacheId?: string;
 }) {
   /*
   * stamped by icestark for recycle when needed.
   */
   element.setAttribute(PREFIX, DYNAMIC);
+  cacheId && element.setAttribute('data-cache', cacheId);
   element.id = id;
-
 
   element.type = module ? 'module' : 'text/javascript';
   element.src = src;
@@ -225,10 +235,12 @@ export function appendExternalScript(asset: string | Asset,
     id,
     root = document.getElementsByTagName('head')[0],
     scriptAttributes = [],
+    cacheId,
   }: {
     id: string;
     root?: HTMLElement | ShadowRoot;
     scriptAttributes?: ScriptAttributes;
+    cacheId?: string;
   }): Promise<void> {
   return new Promise<void>((resolve, reject) => {
     const { type, content, module } = (asset as Asset);
@@ -239,6 +251,7 @@ export function appendExternalScript(asset: string | Asset,
       element.innerHTML = content;
       element.id = id;
       element.setAttribute(PREFIX, DYNAMIC);
+      cacheId && element.setAttribute('data-cache', cacheId);
       module && (element.type = 'module');
       root.appendChild(element);
 
@@ -254,6 +267,7 @@ export function appendExternalScript(asset: string | Asset,
       id,
       src: content || (asset as string),
       scriptAttributes,
+      cacheId,
     });
 
     if (isAssetExist(element, 'script')) {
@@ -653,21 +667,8 @@ export function emptyAssets(
 
 export function checkCacheKey(node: HTMLElement | HTMLLinkElement | HTMLStyleElement | HTMLScriptElement, cacheKey: string|boolean) {
   return (typeof cacheKey === 'boolean' && cacheKey)
-    || !node.getAttribute('cache')
-    || node.getAttribute('cache') === cacheKey;
-}
-
-/**
- * cache all assets loaded by current sub-application
- */
-export function cacheAssets(cacheKey: string): void {
-  const assetsList = getAssetsNode();
-  assetsList.forEach((assetsNode) => {
-    // set cache key if asset attributes without prefix=static and cache
-    if (assetsNode.getAttribute(PREFIX) !== STATIC && !assetsNode.getAttribute('cache')) {
-      assetsNode.setAttribute('cache', cacheKey);
-    }
-  });
+    || !node.getAttribute('data-cache')
+    || node.getAttribute('data-cache') === cacheKey;
 }
 
 /**
@@ -679,9 +680,11 @@ export function cacheAssets(cacheKey: string): void {
 export async function loadAndAppendCssAssets(cssList: Array<Asset | HTMLElement>, {
   cacheCss = false,
   fetch = defaultFetch,
+  cacheId,
 }: {
   cacheCss?: boolean;
   fetch?: Fetch;
+  cacheId?: string;
 }) {
   const cssRoot: HTMLElement = document.getElementsByTagName('head')[0];
 
@@ -704,18 +707,28 @@ export async function loadAndAppendCssAssets(cssList: Array<Asset | HTMLElement>
     // And supposed to be remove from 3.x
     if (!useLinks) {
       return await Promise.all([
-        ...cssContents.map((content, index) => appendCSS(
-          cssRoot,
-          { content, type: AssetTypeEnum.INLINE }, `${PREFIX}-css-${index}`,
-        )),
-        ...cssList.filter((css) => isElement(css)).map((asset, index) => appendCSS(cssRoot, asset, `${PREFIX}-css-${index}`)),
+        ...cssContents.map((content, index) => appendCSS({ content, type: AssetTypeEnum.INLINE },
+          {
+            root: cssRoot,
+            id: `${PREFIX}-css-${index}`,
+            cacheId,
+          })),
+        ...cssList.filter((css) => isElement(css)).map((asset, index) => appendCSS(asset, {
+          root: cssRoot,
+          id: `${PREFIX}-css-${index}`,
+          cacheId,
+        })),
       ]);
     }
   }
 
   // load css content
-  return await Promise.all(
-    cssList.map((asset, index) => appendCSS(cssRoot, asset, `${PREFIX}-css-${index}`)),
+  await Promise.all(
+    cssList.map((asset, index) => appendCSS(asset, {
+      root: cssRoot,
+      id: `${PREFIX}-css-${index}`,
+      cacheId,
+    })),
   );
 }
 
@@ -731,8 +744,10 @@ export function loadAndAppendJsAssets(
   assets: Assets,
   {
     scriptAttributes = [],
+    cacheId,
   }: {
     scriptAttributes?: ScriptAttributes;
+    cacheId?: string;
   },
 ) {
   const jsRoot: HTMLElement = document.getElementsByTagName('head')[0];
@@ -748,6 +763,7 @@ export function loadAndAppendJsAssets(
         root: jsRoot,
         scriptAttributes,
         id: `${PREFIX}-js-${index}`,
+        cacheId,
       }));
     }, Promise.resolve());
   }
@@ -757,6 +773,7 @@ export function loadAndAppendJsAssets(
       root: jsRoot,
       scriptAttributes,
       id: `${PREFIX}-js-${index}`,
+      cacheId,
     })),
   );
 }