+++ packages/sandbox/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.1.5
+
+- [fix] exposes `injection` through constructor.
+
+## 1.1.4
+
+- [fix] simply copy callable funtions's extra properties.
+
 ## 1.1.2
 
 - [fix] hijacked eventListener were not been removed after sandbox unload. ([#295](https://github.com/ice-lab/icestark/issues/295))
@@ -9,6 +17,7 @@
 ## 1.1.1
 
 - [fix] falsy values except `undefined` would be trapped by proxy window. ([#156](https://github.com/ice-lab/icestark/issues/156))
+
 ## 1.1.0
 
 - [feat] mark access to all properties added to local window by using method `getAddedProperties`.